// backends/api-beans/src/.generated/graphql.ts
export type IBaseBeanResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BaseBean'] = IResolversParentTypes['BaseBean']> = {
  __resolveType: TypeResolveFn<'Bean', ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<Maybe<IResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  name?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  originId?: Resolver<Maybe<IResolversTypes['UUID']>, ParentType, ContextType>;
  processingMethodId?: Resolver<Maybe<IResolversTypes['UUID']>, ParentType, ContextType>;
  roastLevelId?: Resolver<Maybe<IResolversTypes['UUID']>, ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
};

export type IBaseBeanProcessingMethodResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BaseBeanProcessingMethod'] = IResolversParentTypes['BaseBeanProcessingMethod']> = {
  __resolveType: TypeResolveFn<'ProcessingMethod', ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<Maybe<IResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  name?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
};

export type IBaseBeanRoastLevelResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BaseBeanRoastLevel'] = IResolversParentTypes['BaseBeanRoastLevel']> = {
  __resolveType: TypeResolveFn<'RoastLevel', ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<Maybe<IResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  name?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
};

export type IBaseFlavorProfileResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BaseFlavorProfile'] = IResolversParentTypes['BaseFlavorProfile']> = {
  __resolveType: TypeResolveFn<'FlavorProfile', ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<Maybe<IResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  label?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  value?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
};

export type IBaseOriginResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BaseOrigin'] = IResolversParentTypes['BaseOrigin']> = {
  __resolveType: TypeResolveFn<'Origin', ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<Maybe<IResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  latitude?: Resolver<Maybe<IResolversTypes['Latitude']>, ParentType, ContextType>;
  longitude?: Resolver<Maybe<IResolversTypes['Longitude']>, ParentType, ContextType>;
  name?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
};

export type IBasePackResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BasePack'] = IResolversParentTypes['BasePack']> = {
  __resolveType: TypeResolveFn<'Pack', ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<Maybe<IResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  name?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
};

export type IBasePurchaseResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BasePurchase'] = IResolversParentTypes['BasePurchase']> = {
  __resolveType: TypeResolveFn<'Purchase', ParentType, ContextType>;
  capacity?: Resolver<IResolversTypes['Int'], ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  packId?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  price?: Resolver<IResolversTypes['Int'], ParentType, ContextType>;
  purchasedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  shopId?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
};

export type IBaseShopResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BaseShop'] = IResolversParentTypes['BaseShop']> = {
  __resolveType: TypeResolveFn<'Shop', ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  name?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
};

export type IBaseTastingResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BaseTasting'] = IResolversParentTypes['BaseTasting']> = {
  __resolveType: TypeResolveFn<'Tasting', ParentType, ContextType>;
  acidity?: Resolver<IResolversTypes['Int'], ParentType, ContextType>;
  conductedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  name?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  richness?: Resolver<IResolversTypes['Int'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
};

export type IBeanResolvers<ContextType = Context, ParentType extends IResolversParentTypes['Bean'] = IResolversParentTypes['Bean']> = {
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<Maybe<IResolversTypes['String']>, ParentType, ContextType>;
  flavorProfiles?: Resolver<Maybe<Array<Maybe<IResolversTypes['FlavorProfile']>>>, ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  name?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  origin?: Resolver<Maybe<IResolversTypes['Origin']>, ParentType, ContextType>;
  originId?: Resolver<Maybe<IResolversTypes['UUID']>, ParentType, ContextType>;
  packs?: Resolver<Maybe<Array<Maybe<IResolversTypes['Pack']>>>, ParentType, ContextType>;
  processingMethod?: Resolver<Maybe<IResolversTypes['ProcessingMethod']>, ParentType, ContextType>;
  processingMethodId?: Resolver<Maybe<IResolversTypes['UUID']>, ParentType, ContextType>;
  roastLevel?: Resolver<Maybe<IResolversTypes['RoastLevel']>, ParentType, ContextType>;
  roastLevelId?: Resolver<Maybe<IResolversTypes['UUID']>, ParentType, ContextType>;
  tastings?: Resolver<Maybe<Array<Maybe<IResolversTypes['Tasting']>>>, ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type IBeanResultsResolvers<ContextType = Context, ParentType extends IResolversParentTypes['BeanResults'] = IResolversParentTypes['BeanResults']> = {
  __resolveType: TypeResolveFn<'Bean' | 'NotFoundError', ParentType, ContextType>;
};

export type IConflictErrorResolvers<ContextType = Context, ParentType extends IResolversParentTypes['ConflictError'] = IResolversParentTypes['ConflictError']> = {
  code?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  message?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type ICreateBeanResultsResolvers<ContextType = Context, ParentType extends IResolversParentTypes['CreateBeanResults'] = IResolversParentTypes['CreateBeanResults']> = {
  __resolveType: TypeResolveFn<'Bean' | 'ConflictError', ParentType, ContextType>;
};

export type ICreateOriginResultsResolvers<ContextType = Context, ParentType extends IResolversParentTypes['CreateOriginResults'] = IResolversParentTypes['CreateOriginResults']> = {
  __resolveType: TypeResolveFn<'ConflictError' | 'Origin', ParentType, ContextType>;
};

export interface IDateTimeIsoScalarConfig extends GraphQLScalarTypeConfig<IResolversTypes['DateTimeISO'], any> {
  name: 'DateTimeISO';
}

export type IFlavorProfileResolvers<ContextType = Context, ParentType extends IResolversParentTypes['FlavorProfile'] = IResolversParentTypes['FlavorProfile']> = {
  beans?: Resolver<Maybe<Array<IResolversTypes['Bean']>>, ParentType, ContextType>;
  createdAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  description?: Resolver<Maybe<IResolversTypes['String']>, ParentType, ContextType>;
  id?: Resolver<IResolversTypes['UUID'], ParentType, ContextType>;
  label?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  updatedAt?: Resolver<IResolversTypes['DateTimeISO'], ParentType, ContextType>;
  value?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
};

export type IGraphQlErrorResolvers<ContextType = Context, ParentType extends IResolversParentTypes['GraphQLError'] = IResolversParentTypes['GraphQLError']> = {
  __resolveType: TypeResolveFn<'ConflictError' | 'NotFoundError', ParentType, ContextType>;
  code?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
  message?: Resolver<IResolversTypes['String'], ParentType, ContextType>;
};
// ...
export type IOriginResultsResolvers<ContextType = Context, ParentType extends IResolversParentTypes['OriginResults'] = IResolversParentTypes['OriginResults']> = {
  __resolveType: TypeResolveFn<'NotFoundError' | 'Origin', ParentType, ContextType>;
};
// ...
export type IResolvers<ContextType = Context> = {
  BaseBean?: IBaseBeanResolvers<ContextType>;
  BaseBeanProcessingMethod?: IBaseBeanProcessingMethodResolvers<ContextType>;
  BaseBeanRoastLevel?: IBaseBeanRoastLevelResolvers<ContextType>;
  BaseFlavorProfile?: IBaseFlavorProfileResolvers<ContextType>;
  BaseOrigin?: IBaseOriginResolvers<ContextType>;
  BasePack?: IBasePackResolvers<ContextType>;
  BasePurchase?: IBasePurchaseResolvers<ContextType>;
  BaseShop?: IBaseShopResolvers<ContextType>;
  BaseTasting?: IBaseTastingResolvers<ContextType>;
  Bean?: IBeanResolvers<ContextType>;
  BeanResults?: IBeanResultsResolvers<ContextType>;
  ConflictError?: IConflictErrorResolvers<ContextType>;
  CreateBeanResults?: ICreateBeanResultsResolvers<ContextType>;
  CreateOriginResults?: ICreateOriginResultsResolvers<ContextType>;
  DateTimeISO?: GraphQLScalarType;
  FlavorProfile?: IFlavorProfileResolvers<ContextType>;
  GraphQLError?: IGraphQlErrorResolvers<ContextType>;
  Latitude?: GraphQLScalarType;
  Longitude?: GraphQLScalarType;
  Mutation?: IMutationResolvers<ContextType>;
  NotFoundError?: INotFoundErrorResolvers<ContextType>;
  Origin?: IOriginResolvers<ContextType>;
  OriginResults?: IOriginResultsResolvers<ContextType>;
  Pack?: IPackResolvers<ContextType>;
  ProcessingMethod?: IProcessingMethodResolvers<ContextType>;
  Purchase?: IPurchaseResolvers<ContextType>;
  Query?: IQueryResolvers<ContextType>;
  RoastLevel?: IRoastLevelResolvers<ContextType>;
  Shop?: IShopResolvers<ContextType>;
  Tasting?: ITastingResolvers<ContextType>;
  URL?: GraphQLScalarType;
  UUID?: GraphQLScalarType;
};